\docType{package}
\name{genesSelector-package}
\alias{genesSelector-package}
\title{Using pre-processed DNA chip data, this package enable the user to
use a GUI to select genes. Base criteria include fold & p-value. Consult
the pdf help document for details.}
\description{
  Using pre-processed DNA chip data, this package enable
  the user to use a GUI to select genes. Base criteria
  include fold & p-value. Consult the pdf help document for
  details.
}

