\docType{package}
\name{genesSelector}
\alias{genesSelector}
\alias{genesSelector-package}
\title{A GUI to facilitate gene selection based on p-value and folds criteria from DNA chip experiments}
\description{
  A GUI to facilitate gene selection based on p-value and
  folds criteria from DNA chip experiments
}
\details{
  Using pre-processed DNA chip data, this package enable
  the user to use a GUI to select genes. Base criteria
  include fold & p-value. Consult the pdf help document for
  details.

  The main function of this package is
  \code{\link{genesSelectorGUI}}
}

