\name{genesSelectorGUI}
\alias{genesSelectorGUI}
\title{Launch the genes selector GUI}
\usage{
  genesSelectorGUI()
}
\description{
  This function start a new Graphical User Interface for
  the genes selector. Make sure you have not any instance
  of the GUI running before calling this function. The full
  documentation on this tool can be acessed through the
  help manual, located in the "doc" subfolder of the
  package installation folder.
}
\keyword{gene,}
\keyword{GUI,}
\keyword{selection}

