\name{setCurrentStep}
\alias{setCurrentStep}
\title{Change the current step}
\usage{
  setCurrentStep(step)
}
\arguments{
  \item{step}{integer, currently between 0 (file loading
  screen) and 4}
}
\description{
  Desactivate widgets and change currentStep global
  variable. Does not change focus or hide or unhide
  windows.
}

