\name{saveResults}
\alias{saveResults}
\title{Save current results in a folder}
\usage{
  saveResults(saveFolder)
}
\arguments{
  \item{saveFolder}{the path to the folder which will
  contain the results}
}
\description{
  This function uses global variables created by the
  package (thus need only one parameter) Create the
  following subfolders :
}
\details{
  Create the following files :
}

